@@ -5,21 +5,26 @@
     "keywords": ["framework", "laravel"],
     "license": "MIT",
     "require": {
-        "php": "^7.3|^8.0",
+        "php": "^8.3",
+        "algolia/algoliasearch-client-php": "^3.3",
         "arcanedev/log-viewer": "^8.3",
         "barryvdh/laravel-debugbar": "^3.6",
+        "dusterio/link-preview": "^1.2",
         "fruitcake/laravel-cors": "^2.0",
         "guzzlehttp/guzzle": "^7.0.1",
         "laravel/framework": "^8.75",
         "laravel/sanctum": "^2.11",
+        "laravel/scout": "^9.4",
         "laravel/tinker": "^2.5",
         "laravel/ui": "^3.4",
         "predis/predis": "^1.1",
+        "stevebauman/purify": "^4.0",
         "ylsideas/feature-flags": "^1.5"
     },
     "require-dev": {
         "facade/ignition": "^2.5",
         "fakerphp/faker": "^1.9.1",
+        "laravel/pint": "^1.24",
         "laravel/sail": "^1.0.1",
         "mockery/mockery": "^1.4.4",
         "nunomaduro/collision": "^5.10",
