@@ -2,11 +2,7 @@
 
 namespace App\Events;
 
-use Illuminate\Broadcasting\Channel;
 use Illuminate\Broadcasting\InteractsWithSockets;
-use Illuminate\Broadcasting\PresenceChannel;
-use Illuminate\Broadcasting\PrivateChannel;
-use Illuminate\Contracts\Broadcasting\ShouldBroadcast;
 use Illuminate\Foundation\Events\Dispatchable;
 use Illuminate\Queue\SerializesModels;
 
@@ -14,15 +10,20 @@ class ThreadReceivedNewReply
 {
     use Dispatchable, InteractsWithSockets, SerializesModels;
 
+    public $thread;
+
     public $reply;
 
     /**
      * Create a new event instance.
      *
+     * @param  \App\Thread  $thread
+     * @param  \App\Reply  $reply
      * @return void
      */
-    public function __construct($reply)
+    public function __construct($thread, $reply)
     {
+        $this->thread = $thread;
         $this->reply = $reply;
     }
 }
