@@ -4,7 +4,4 @@ namespace App\Exceptions;
 
 use Exception;
 
-class ThrottleException extends Exception
-{
-
-}
+class ThrottleException extends Exception {}
