@@ -4,18 +4,19 @@ namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
 use App\Models\User;
-use Illuminate\Http\Request;
 
 class RegisterConfirmationController extends Controller
 {
     public function index()
     {
         $user = User::where('confirmation_token', request('token'))->first();
-        
-        if (!$user) { return redirect(route('threads'))->with('flash', 'Unknown token'); }
+
+        if (! $user) {
+            return redirect(route('threads'))->with('flash', 'Unknown token');
+        }
 
         $user->confirm();
-        
+
         return redirect(route('threads'))
             ->with('flash', 'Your account is now confirmed! You may post to the forum.');
     }
