@@ -3,16 +3,20 @@
 namespace App\Http\Controllers;
 
 use App\Models\Reply;
-use Illuminate\Http\Request;
 
 class BestRepliesController extends Controller
 {
+    /**
+     * Store a newly created resource in storage.
+     *
+     * @return void
+     *
+     * @throws \Illuminate\Auth\Access\AuthorizationException
+     */
     public function store(Reply $reply)
     {
-        $this->authorize('update', $reply->thread);
-        // abort_if($reply->thread->user_id !== auth()->id, 401);
+        // $this->authorize('update', $reply->thread);
 
-        // $reply->thread->update(['best_reply_id' => $reply->id]);
         $reply->thread->markBestReply($reply);
     }
 }
