@@ -2,16 +2,17 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Channel;
 use App\Models\Thread;
 
 class ThreadSubscriptionsController extends Controller
 {
-    public function store($channelId, Thread $thread)
+    public function store(Channel $channel, Thread $thread)
     {
         $thread->subscribe();
     }
 
-    public function destroy($channelId, Thread $thread)
+    public function destroy(Channel $channel, Thread $thread)
     {
         $thread->unsubscribe();
     }
