@@ -10,16 +10,15 @@ class RedirectIfEmailNotConfirmed
     /**
      * Handle an incoming request.
      *
-     * @param  \Illuminate\Http\Request  $request
      * @param  \Closure(\Illuminate\Http\Request): (\Illuminate\Http\Response|\Illuminate\Http\RedirectResponse)  $next
      * @return \Illuminate\Http\Response|\Illuminate\Http\RedirectResponse
      */
     public function handle(Request $request, Closure $next)
     {
-        if (!$request->user()->confirmed) {
+        if (! $request->user()->confirmed) {
             return redirect('/threads')
                 ->with('flash', 'You must first confirm your email address.');
-        };
+        }
 
         return $next($request);
     }
