@@ -40,7 +40,7 @@ class CreatePostRequest extends FormRequest
     public function rules()
     {
         return [
-            'body' => 'required|spamfree'
+            'body' => 'required|string|spamfree',
         ];
     }
 }
