@@ -2,15 +2,11 @@
 
 namespace App\Inspections;
 
-use Exception;
-use App\Inspections\KeyHeldDown;
-use App\Inspections\InvalidKeywords;
-
 class Spam
 {
     protected $inspections = [
         InvalidKeywords::class,
-        KeyHeldDown::class
+        KeyHeldDown::class,
     ];
 
     public function detect($body)
