@@ -2,20 +2,16 @@
 
 namespace App\Listeners;
 
-use App\Models\User;
 use App\Events\ThreadReceivedNewReply;
+use App\Models\User;
 use App\Notifications\YouWereMentioned;
-use Illuminate\Queue\InteractsWithQueue;
-use Illuminate\Contracts\Queue\ShouldQueue;
+use Illuminate\Support\Facades\Notification;
 
 class NotifyMentionedUsers
 {
     public function handle(ThreadReceivedNewReply $event)
     {
-        User::whereIn('name', $event->reply->mentionedUsers())
-            ->get()
-            ->each(function ($user) use ($event) {
-                $user->notify(new YouWereMentioned($event->reply));
-            });
+        $users = User::whereIn('name', $event->reply->mentionedUsers())->get();
+        Notification::send($users, new YouWereMentioned($event->reply));
     }
 }
