@@ -3,8 +3,6 @@
 namespace App\Listeners;
 
 use App\Events\ThreadReceivedNewReply;
-use Illuminate\Contracts\Queue\ShouldQueue;
-use Illuminate\Queue\InteractsWithQueue;
 
 class NotifySubscribers
 {
@@ -26,7 +24,7 @@ class NotifySubscribers
      */
     public function handle(ThreadReceivedNewReply $event)
     {
-        $event->reply->thread->subscriptions
+        $event->thread->subscriptions
             ->where('user_id', '!=', $event->reply->user_id)
             ->each
             ->notify($event->reply);
