@@ -2,19 +2,30 @@
 
 namespace App\Models;
 
-use App\Models\User;
-use ReflectionClass;
-use App\Models\Reply;
-use App\Service\Visits;
-use Illuminate\Support\Str;
-use App\Traits\RecordsActivity;
 use App\Events\ThreadReceivedNewReply;
-use Illuminate\Database\Eloquent\Model;
+use App\Service\Visits;
+use App\Traits\RecordsActivity;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Str;
+use Laravel\Scout\Searchable;
 
 class Thread extends Model
 {
-    use HasFactory, RecordsActivity;
+    use HasFactory, RecordsActivity, Searchable;
+
+    protected static function boot()
+    {
+        parent::boot();
+
+        static::addGlobalScope('replyCount', function ($builder) {
+            $builder->withCount('replies');
+        });
+
+        self::deleting(function ($thread) {
+            $thread->replies->each->delete();
+        });
+    }
 
     protected $guarded = [];
 
@@ -22,45 +33,19 @@ class Thread extends Model
 
     protected $appends = ['isSubscribedTo'];
 
+    protected $casts = [
+        'locked' => 'boolean',
+    ];
+
     public function path()
     {
-        return '/threads/' . $this->channel->slug . '/' . $this->slug;
-    }
-
-    protected static function boot()
-    {
-        parent::boot();
-
-        // static::addGlobalScope('replyCount', function ($builder)
-        // {
-        //     $builder->withCount('replies');
-        // });
-
-        static::deleting(function ($thread)
-        {
-            // $thread->replies()->each()->delete();
-            $thread->replies()->each(function ($reply) {
-                $reply->delete();
-            });
-        });
-
-        static::created(function ($thread)
-        {
-            $thread->update(['slug' => $thread->title]);
-        });
+        return "/threads/{$this->channel->slug}/$this->slug";
     }
 
     public function replies()
     {
         return $this->hasMany(Reply::class);
-            // ->withCount('favorites')
-            // ->with('creator');
     }
-
-    // public function getReplyCountAttribute()
-    // {
-    //     return $this->replies()->count();
-    // }
 
     public function channel()
     {
@@ -72,15 +57,18 @@ class Thread extends Model
         return $this->belongsTo(User::class, 'user_id');
     }
 
+    public function addReplies($replies)
+    {
+        foreach ($replies as $reply) {
+            $this->addReply($reply);
+        }
+    }
+
     public function addReply($reply)
     {
-        // (new Spam)->detect($reply->body);
-// dd($reply);
         $reply = $this->replies()->create($reply);
 
-        event(new ThreadReceivedNewReply($reply));
-
-        // $this->notifySubscribers($reply);
+        event(new ThreadReceivedNewReply($this, $reply));
 
         return $reply;
     }
@@ -92,16 +80,19 @@ class Thread extends Model
 
     public function subscribe($userId = null)
     {
-        $this->subscriptions()->create([
-            'user_id' => $userId ?: auth()->id()
-        ]);
+        $this->subscriptions()
+            ->create([
+                'user_id' => $userId ?: auth()->id(),
+            ]);
+
+        return $this;
     }
 
     public function unsubscribe($userId = null)
     {
         $this->subscriptions()
-        ->where('user_id', $userId ?: auth()->id())
-        ->delete();
+            ->where('user_id', $userId ?: auth()->id())
+            ->delete();
     }
 
     public function subscriptions()
@@ -109,25 +100,55 @@ class Thread extends Model
         return $this->hasMany(ThreadSubscription::class);
     }
 
-    public function getIsSubscribedToAttribute()
+    protected function isSubscribedTo($userId = null)
     {
-        return $this->subscriptions()
-            ->where('user_id', auth()->id())
-            ->exists();
+        return $this->subscriptions()->where('user_id', $userId ?: auth()->id())->exists();
     }
 
-    public function hasUpdatesFor($user)
+    public function getIsSubscribedToAttribute()
     {
-        // $key = sprintf("users.%s.visits.%s", auth()->id(), $this->id);
-        $key = $user->visitedThreadCacheKey($this);
+        return $this->isSubscribedTo();
+    }
+
+    public function hasUpdates($user)
+    {
+        if (! \Auth::check()) {
+            return true;
+        }
+
+        $key = \Auth::user()->visitedThreadCacheKey($this);
 
         return $this->updated_at > cache($key);
     }
 
-    // public function visits()
-    // {
-    //     return new Visits($this);
-    // }
+    /**
+     * Lock a thread from receiving replies.
+     *
+     * @return bool
+     */
+    public function lock()
+    {
+        return $this->update([
+            'is_locked' => true,
+        ]);
+    }
+
+    /**
+     * Unlock a thread from receiving replies.
+     *
+     * @return bool
+     */
+    public function unlock()
+    {
+        return $this->update([
+            'is_locked' => false,
+        ]);
+    }
+
+    public function visits()
+    {
+        return new Visits($this);
+    }
 
     public function getRouteKeyName()
     {
@@ -139,15 +160,33 @@ class Thread extends Model
         $slug = Str::slug($value);
 
         if (static::whereSlug($slug)->exists()) {
-            $slug = "{$slug}-" . $this->id;
+            $slug = $this->incrementSlug($slug);
         }
 
         $this->attributes['slug'] = $slug;
     }
 
+    /**
+     * Increment a slug's suffix.
+     *
+     * @param  string  $slug
+     * @return string
+     */
+    protected function incrementSlug($slug, $count = 2)
+    {
+        $original = $slug;
+
+        while (static::whereSlug($slug)->exists()) {
+            $slug = $original.'-'.$count;
+        }
+
+        return $slug;
+    }
+
     public function markBestReply(Reply $reply)
     {
-        // $this->best_reply_id = $reply->id;
-        $this->update(['best_reply_id' => $reply->id]);
+        $reply->thread->update([
+            'best_reply_id' => $reply->id,
+        ]);
     }
 }
