@@ -2,6 +2,7 @@
 
 namespace App\Models;
 
+use App\Notifications\ThreadWasUpdated;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
@@ -10,7 +11,7 @@ class ThreadSubscription extends Model
     use HasFactory;
 
     protected $guarded = [];
-    
+
     /**
      * Get the user that owns the ThreadSubscription
      *
@@ -20,7 +21,7 @@ class ThreadSubscription extends Model
     {
         return $this->belongsTo(User::class);
     }
-    
+
     public function thread()
     {
         return $this->belongsTo(Thread::class);
@@ -28,6 +29,6 @@ class ThreadSubscription extends Model
 
     public function notify($reply)
     {
-        return $this->user->notify(new ThreadWasUpdated($this->thread, $reply));
+        $this->user->notify(new ThreadWasUpdated($this->thread, $reply));
     }
 }
