@@ -3,12 +3,12 @@
 namespace App\Providers;
 
 use App\Models\Channel;
+use Barryvdh\Debugbar\ServiceProvider as BarryvdhServiceProvider;
 use Illuminate\Pagination\Paginator;
 use Illuminate\Support\Facades\Cache;
-use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\ServiceProvider;
 use YlsIdeas\FeatureFlags\Facades\Features;
-use Barryvdh\Debugbar\ServiceProvider as BarryvdhServiceProvider;
 
 class AppServiceProvider extends ServiceProvider
 {
@@ -34,9 +34,11 @@ class AppServiceProvider extends ServiceProvider
         Paginator::useBootstrap();
 
         view()->composer('*', function ($view) {
-            $channels = Cache::rememberForever('channels', function () {
-                return Channel::all();
-            });
+            // $channels = Cache::rememberForever('channels', function () {
+            //     return Channel::all();
+            // });
+
+            $channels = Channel::all();
 
             $view->with('channels', $channels);
         });
