@@ -19,17 +19,20 @@
     <!-- Styles -->
     <link rel="stylesheet" href="{{ mix('css/app.css') }}" />
     <script defer src="{{ mix('js/app.js') }}"></script>
+    {{-- <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/trix/1.3.1/trix.css" integrity="sha512-CWdvnJD7uGtuypLLe5rLU3eUAkbzBR3Bm1SFPEaRfvXXI2v2H5Y0057EMTzNuGGRIznt8+128QIDQ8RqmHbAdg==" crossorigin="anonymous" referrerpolicy="no-referrer" /> --}}
 
     <!-- include libraries(jQuery, bootstrap) -->
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/css/bootstrap.min.css">
     <script src="https://code.jquery.com/jquery-3.5.1.min.js"></script>
-    
+
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"></script>
     <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js"></script>
     <script src="https://kit.fontawesome.com/8245952001.js" crossorigin="anonymous"></script>
 
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/Caret.js/0.3.1/jquery.caret.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/at.js/1.5.0/js/jquery.atwho.min.js"></script>
+    {{-- <script src="https://cdnjs.cloudflare.com/ajax/libs/Caret.js/0.3.1/jquery.caret.min.js"></script> --}}
+    {{-- <script src="https://cdnjs.cloudflare.com/ajax/libs/at.js/1.5.0/js/jquery.atwho.min.js"></script> --}}
+
+    <script src="https://www.google.com/recaptcha/api.js" async defer></script>
 
     <script>
         window.App = {!! json_encode([
@@ -41,6 +44,10 @@
 
     <style>
         [v-cloak] { display: none; }
+        #threadLink:link { text-decoration: none; }
+        #threadLink:visited { text-decoration: none; }
+        #threadLink:hover { text-decoration: none; }
+        #threadLink:active { text-decoration: none; }
     </style>
 
     @yield('header')
@@ -52,7 +59,7 @@
         <main class="py-4">
             @yield('content')
         </main>
-        
+
         <flash message="{{ session('flash') }}"></flash>
         {{-- <flash message="temporary"></flash> --}}
     </div>
