@@ -8,6 +8,6 @@
     @endslot
 
     @slot('body')
-        {{ $activity->subject->favorited->body }}
+        {!! $activity->subject->favorited->body !!}
     @endslot
 @endcomponent
