@@ -1,6 +1,6 @@
 @component('profiles.activities.activity')
     @slot('heading')
-        {{ $profileUser->name }} replied to <a href="{{ $activity->subject->thread->path() ?? '-' }}">{{ $activity->subject->thread->title ?? '-' }}</a>
+        {{ $profileUser->name }} replied to <a href="{{ $activity->subject->thread->path() ?? '-' }}">{{ $activity->subject->thread->subject ?? '-' }}</a>
     @endslot
 
     @slot('time')
@@ -8,6 +8,6 @@
     @endslot
 
     @slot('body')
-        {{ $activity->subject->body }}
+        {!! $activity->subject->body !!}
     @endslot
 @endcomponent
